+++ src/local-env-actions.js
@@ -5,18 +5,19 @@ import {
   ENV_PATH,
   SNOWDOG_COMPONENTS,
   ALPACA_THEME_DIR,
-  BASE_PATH,
+  BASE_THEME_PATH,
   MAGENTO_CHECKOUT_STYLES
 } from './constants/constants.js'
 import {
   addFilesFromDir,
   addFilesFromTemplate,
   replacePhraseInAll,
-  prependImport
+  prependImport,
+  replacePhrase
 } from './local-env-helper.js'
 
 // SETING UP SNOWDOG_COMPONENTS CONFIG FILES
-export async function setupComponentsConfigFiles(themeName, fullThemeName) {
+export async function setupComponentsConfigFiles(themeName, fullThemeName, vendor) {
   const componentFilesToUpdate = [
     {
       name: 'gulpfile.mjs',
@@ -30,16 +31,16 @@ export async function setupComponentsConfigFiles(themeName, fullThemeName) {
     }
   ]
   try {
-    await addFilesFromDir(ENV_PATH.ALPACA_COMPONENTS_DIR, themeName, '.lock|.md', SNOWDOG_COMPONENTS)
-    await addFilesFromTemplate(ENV_PATH.TEMPLATES_COMPONENTS_CONFIG_DIR, `${BASE_PATH}${themeName}${SNOWDOG_COMPONENTS}`)
-    await replacePhraseInAll(componentFilesToUpdate, `${BASE_PATH}${themeName}${SNOWDOG_COMPONENTS}`)
+    await addFilesFromDir(ENV_PATH.ALPACA_COMPONENTS_DIR, themeName, '.lock|.md', vendor, SNOWDOG_COMPONENTS)
+    await addFilesFromTemplate(ENV_PATH.TEMPLATES_COMPONENTS_CONFIG_DIR, `${BASE_THEME_PATH}${vendor}/${themeName}${SNOWDOG_COMPONENTS}`)
+    await replacePhraseInAll(componentFilesToUpdate, `${BASE_THEME_PATH}${vendor}/${themeName}${SNOWDOG_COMPONENTS}`)
   } catch (error) {
     console.log(`\n${error}`)
   }
 }
 
 // SETTING UP THEME LEVEL BASE CONFIG FILES
-export async function setupThemeConfigFiles(themeName, fullThemeName) {
+export async function setupThemeConfigFiles(themeName, fullThemeName, vendor) {
   const lineToAddParentTag = 2
   const parentTag = '    <parent>Snowdog/alpaca</parent>'
   const ignoredFiles = '.lock|.md|now|LICENSE|composer'
@@ -61,11 +62,11 @@ export async function setupThemeConfigFiles(themeName, fullThemeName) {
     }
   ]
 
-  await addFilesFromDir(ALPACA_THEME_DIR, themeName, ignoredFiles)
-  await addFilesFromTemplate(ENV_PATH.TEMPLATES_THEME_DIR, `${BASE_PATH}${themeName}`)
-  await replacePhraseInAll(themeFilesToUpdate, `${BASE_PATH}${themeName}`)
+  await addFilesFromDir(ALPACA_THEME_DIR, themeName, ignoredFiles, vendor)
+  await addFilesFromTemplate(ENV_PATH.TEMPLATES_THEME_DIR, `${BASE_THEME_PATH}${vendor}/${themeName}`)
+  await replacePhraseInAll(themeFilesToUpdate, `${BASE_THEME_PATH}${vendor}/${themeName}`)
   await prependImport(
-    `${BASE_PATH}${themeName}/theme.xml`,
+    `${BASE_THEME_PATH}${vendor}/${themeName}/theme.xml`,
     parentTag,
     themeName,
     lineToAddParentTag
@@ -73,7 +74,7 @@ export async function setupThemeConfigFiles(themeName, fullThemeName) {
 }
 
 // CONFIGURING FRONTOOLS
-export async function setupFrontoolsConfigFiles(themeName) {
+export async function setupFrontoolsConfigFiles(themeName, vendor) {
   const frontoolsFilesToUpdate = [
     {
       name: 'browser-sync.json',
@@ -89,29 +90,30 @@ export async function setupFrontoolsConfigFiles(themeName) {
 
   await addFilesFromTemplate(ENV_PATH.TEMPLATES_FRONTOOLS_DIR, ENV_PATH.DEV_FRONTOOLS_CONFIG_DIR)
   await replacePhraseInAll(frontoolsFilesToUpdate, ENV_PATH.DEV_FRONTOOLS_CONFIG_DIR)
+  await replacePhrase('dev/tools/frontools/config/themes.json', 'YOUR_VENDOR', vendor)
 }
 
 // ADDING ALPACA BASE STYLES
-export async function addBaseStyles(themeName) {
-  const docsPath = `${BASE_PATH}${themeName}${ENV_PATH.COMPONENT_DOCS_STYLES_DIR}`
+export async function addBaseStyles(themeName, vendor) {
+  const docsPath = `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.COMPONENT_DOCS_STYLES_DIR}`
   const docsText = VARIABLES_IMPORT_PATHS.COMMENT + VARIABLES_IMPORT_PATHS.DOCS
-  const chechoutPath = `${BASE_PATH}${themeName}${MAGENTO_CHECKOUT_STYLES}/checkout.scss`
+  const chechoutPath = `${BASE_THEME_PATH}${vendor}/${themeName}${MAGENTO_CHECKOUT_STYLES}/checkout.scss`
   const checkoutText = VARIABLES_IMPORT_PATHS.COMMENT + VARIABLES_IMPORT_PATHS.CHECKOUT
-  const themeLevelStylesPath = `${BASE_PATH}${themeName}/styles`
+  const themeLevelStylesPath = `${BASE_THEME_PATH}${vendor}/${themeName}/styles`
   const themeLevelStylesText = VARIABLES_IMPORT_PATHS.COMMENT + VARIABLES_IMPORT_PATHS.MAIN
 
   // CREATING CHILD THEME VARIABLES
   await addFilesFromTemplate(
     ENV_PATH.TEMPLATES_COMPONENTS_BASE_DIR,
-    `${BASE_PATH}${themeName}${ENV_PATH.COMPONENT_VARIABLES_DIR}`
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.COMPONENT_VARIABLES_DIR}`
   )
   await rename(
-    `${BASE_PATH}${themeName}${ENV_PATH.COMPONENT_VARIABLES_FILE}`,
-    `${BASE_PATH}${themeName}${ENV_PATH.COMPONENT_VARIABLES_DIR}/_${themeName}-variables.scss`
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.COMPONENT_VARIABLES_FILE}`,
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.COMPONENT_VARIABLES_DIR}/_${themeName}-variables.scss`
   )
 
   // IMPORTING ALPACA COMPONENTS DOCS STYLES
-  await addFilesFromDir(ENV_PATH.ALPACA_COMPONENTS_DOCS_STYLES_DIR, themeName, '_', ENV_PATH.COMPONENT_DOCS_STYLES_DIR)
+  await addFilesFromDir(ENV_PATH.ALPACA_COMPONENTS_DOCS_STYLES_DIR, themeName, '_', vendor, ENV_PATH.COMPONENT_DOCS_STYLES_DIR)
 
   const docsFilesNames = await listFiles(docsPath)
 
@@ -120,11 +122,11 @@ export async function addBaseStyles(themeName) {
   })
 
   // IMPORTING MAGENTO CHECKOUT STYLES
-  await addFilesFromDir(ENV_PATH.ALPACA_MAGENTO_CHECKOUT_STYLES_DIR, themeName, null, MAGENTO_CHECKOUT_STYLES)
+  await addFilesFromDir(ENV_PATH.ALPACA_MAGENTO_CHECKOUT_STYLES_DIR, themeName, null, vendor, MAGENTO_CHECKOUT_STYLES)
   await prependImport(chechoutPath, checkoutText, themeName, null, 'variables', 'YOUR_THEME_NAME')
 
   // CREATING THEME LEVEL STYLES
-  await addFilesFromDir(ENV_PATH.ALPACA_STYLES_DIR, themeName, null, '/styles')
+  await addFilesFromDir(ENV_PATH.ALPACA_STYLES_DIR, themeName, null, vendor, '/styles')
 
   const themeLevelStyles = await listFiles(themeLevelStylesPath)
 
@@ -141,7 +143,7 @@ export async function addBaseStyles(themeName) {
 }
 
 // ADDING EXEMPLARY BUTTON STYLES AND CRITICAL/NON-CRITICAL IMPORTS
-export async function addExemplaryStyles(themeName) {
+export async function addExemplaryStyles(themeName, vendor) {
   const exemplaryFilesToUpdate = [
     {
       name: '_button-extend.scss',
@@ -166,16 +168,16 @@ export async function addExemplaryStyles(themeName) {
   // ADDING BUTTON STYLES
   await addFilesFromTemplate(
     ENV_PATH.TEMPLATES_COMPONENTS_EXEMPLARY_DIR,
-    `${BASE_PATH}${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button`
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button`
   )
-  await replacePhraseInAll(exemplaryFilesToUpdate, `${BASE_PATH}${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button`)
+  await replacePhraseInAll(exemplaryFilesToUpdate, `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button`)
   await rename(
-    `${BASE_PATH}${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/button.scss`,
-    `${BASE_PATH}${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/_${themeName}-button.scss`
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/button.scss`,
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/_${themeName}-button.scss`
   )
   await rename(
-    `${BASE_PATH}${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/button-variables.scss`,
-    `${BASE_PATH}${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/_${themeName}-button-variables.scss`
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/button-variables.scss`,
+    `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_MOLECULES_DIR}/button/_${themeName}-button-variables.scss`
   )
 
   // IMPORTING CRITICAL AND NON CRITICAL STYLES
@@ -183,7 +185,8 @@ export async function addExemplaryStyles(themeName) {
     ENV_PATH.ALPACA_COMPONENTS_STYLES_DIR,
     themeName,
     'mixins|-extends|_checkout',
+    vendor,
     ENV_PATH.SNOWDOG_COMPONENTS_STYLES_DIR
   )
-  await replacePhraseInAll(criticalStylesToUpdate, `${BASE_PATH}${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_STYLES_DIR}`)
+  await replacePhraseInAll(criticalStylesToUpdate, `${BASE_THEME_PATH}${vendor}/${themeName}${ENV_PATH.SNOWDOG_COMPONENTS_STYLES_DIR}`)
 }