+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+- Add mising styles imports, rework addFilesFromDir method (#10)
 
 ## [Released]
 ## [1.0.0-rc.2] - 25.03.2022