+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased]
 - Add custom vendor name functionality (#2)
 - Add mising styles imports, rework addFilesFromDir method (#10)
+- Import child theme variables before default ones (#13)
+- Add Alpaca Docs info
 
 ## [Released]
 ## [1.0.0-rc.2] - 25.03.2022